feat(fullname): greeter and unlock dialog support fullname
- 登录以及解锁框支持可配置FullName显示功能

Related #40540

+++ lib/common/virtual-keyboard.cpp
@@ -49,7 +49,7 @@ VirtualKeyboard *VirtualKeyboard::instance()
 
 VirtualKeyboard::~VirtualKeyboard()
 {
-    if (m_process->state() != QProcess::NotRunning)
+    if (m_process && m_process->state() != QProcess::NotRunning)
     {
         m_process->terminate();
         m_process->waitForFinished();
